# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/scrap":{"post":{"operationId":"ScrapController_Scrap","parameters":[],"responses":{"201":{"description":""}}}},"/file/upload":{"post":{"operationId":"FilesController_upload","summary":"File or Image Upload Dynamic Bucket","parameters":[{"name":"Authorization","in":"header","description":"Bearer Token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/FileUploadDTO"}}}},"responses":{"201":{"description":"Return upload information."},"400":{"description":"Invalid data"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["File Upload"],"security":[{"bearer":[]}]}},"/properties":{"post":{"operationId":"PropertiesController_create","summary":"Create a new record","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePropertyDto"}}}},"responses":{"201":{"description":"Return new record."},"400":{"description":"Invalid data"},"405":{"description":"Method not allowed"},"406":{"description":"Record already exist"},"500":{"description":"Server Error!"}},"tags":["Properties"]},"get":{"operationId":"PropertiesController_findAll","summary":"Fetches records","parameters":[],"responses":{"200":{"description":"Returns list of records."},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Properties"]}},"/properties/{id}":{"get":{"operationId":"PropertiesController_findOne","summary":"Get a record by id","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns record."},"404":{"description":"Record Not found."},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Properties"]},"patch":{"operationId":"PropertiesController_update","summary":"Update record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePropertyDto"}}}},"responses":{"200":{"description":"Return record"},"404":{"description":"Record not found"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Properties"]},"put":{"operationId":"PropertiesController_updatePut","summary":"Update record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePropertyDto"}}}},"responses":{"200":{"description":"Return record"},"404":{"description":"Record not found"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Properties"]},"delete":{"operationId":"PropertiesController_remove","summary":"delete record","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete record"},"404":{"description":"Record not found"},"405":{"description":"Method not allowed"},"500":{"description":"Server Error!"}},"tags":["Properties"]}}},"info":{"title":"Api","description":"This api will help clients to store their data.","version":"1.0","contact":{}},"tags":[{"name":"Api","description":""}],"servers":[],"components":{"schemas":{"FileUploadDTO":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]},"CreatePropertyDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"country":{"type":"string"},"zipcode":{"type":"string"},"phone":{"type":"string"},"capacity":{"type":"number"},"files":{"type":"array","items":{"type":"string"}},"lat":{"type":"number"},"lng":{"type":"number"}}},"UpdatePropertyDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"address":{"type":"string"},"city":{"type":"string"},"state":{"type":"string"},"country":{"type":"string"},"zipcode":{"type":"string"},"phone":{"type":"string"},"capacity":{"type":"number"},"files":{"type":"array","items":{"type":"string"}},"lat":{"type":"number"},"lng":{"type":"number"}}}}}}